MCP/requirements.txt
fastmcp>=1.0.0
mcp>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
